chore: merge test_array_symbols into array_expressions
Filename test_array_symbols didn't make much sense anymore
